Added increased error handling in test suite
Set so that if json is poorly formatted it will not break the whole thing, instead the log will show the file that failed and carry on. Additionally added a check for None in the file and replaced it with null for proper json formatting. Reduced scope of tests to single day